auth: move auth funcs for consistency
